Changed structure.
