+++ Mindbox.I18n.Abstractions/Mindbox.I18n.Abstractions.csproj
@@ -15,6 +15,5 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.0" />
-    <PackageReference Include="Mindbox.Quokka.Abstractions" Version="4.10.0" />
   </ItemGroup>
 </Project>
